GUI done, add pitch bend range control

+++ Source/Constants.h
@@ -9,7 +9,7 @@
 */
 
 #pragma once
-const float WINDOW_LENGTH_MIN = 2400.0; // 5ms if sample rate is 48kHz
-const float WINDOW_LENGTH_MAX = 20000.0;
+const int WINDOW_LENGTH_MIN = 2400; // 5ms if sample rate is 48kHz
+const int WINDOW_LENGTH_MAX = 20000;
 const int NUM_VOICES = 16;
 const double root = std::exp(std::log(2) / 1200);
